add xlsx_remove_sheet_protection.sh
This script is intended to remove sheet-protection from MS Excel files
without knowing the password. This is "ok", since this feature is not
meant to keep anything secret anyway (nothing is encrypted), but is only
meant to avoid accidental overwrites.
More specifically, it was written because (at least at the time of writing
this) libreoffice could not remove the protection without the password, or
in this case when an empty password was provided for the specific files I
had to deal with - and also because protecting the sheets meant that
formulae I wanted to see were hidden as well.